+++ gen/properties.nim
@@ -11,27 +11,118 @@ type
     CAT, CCC, BI, QC
 
 const
-  categoryNames* = [
-    "Cn", "Lu", "Ll", "Lt", "Mn", "Mc", "Me", "Nd",
-    "Nl", "No", "Zs", "Zl", "Zp", "Cc", "Cf", "Cs", "Co", "Cn", "Lm",
-    "Lo", "Pc", "Pd", "Ps", "Pe", "Pi", "Pf", "Po", "Sm", "Sc", "Sk",
-    "So"
-  ]
+  ctgLm = 0x01
+  ctgLo = 0x02
+  ctgLu = 0x04
+  ctgLl = 0x08
+  ctgLt = 0x10
+  ctgMn = 0x20
+  ctgMc = 0x40
+  ctgMe = 0x80
+  ctgNd = 0x100
+  ctgNl = 0x200
+  ctgNo = 0x400
+  ctgZs = 0x800
+  ctgZl = 0x1000
+  ctgZp = 0x2000
+  ctgCc = 0x4000
+  ctgCf = 0x8000
+  ctgCs = 0x10000
+  ctgCo = 0x20000
+  ctgCn = 0x40000
+  ctgPc = 0x80000
+  ctgPd = 0x100000
+  ctgPs = 0x200000
+  ctgPe = 0x400000
+  ctgPi = 0x800000
+  ctgPf = 0x1000000
+  ctgPo = 0x2000000
+  ctgSm = 0x4000000
+  ctgSc = 0x8000000
+  ctgSk = 0x10000000
+  ctgSo = 0x20000000
+
   bidirectionalNames* = [
     "L", "LRE", "LRO", "R", "AL", "RLE", "RLO",
     "PDF", "EN", "ES", "ET", "AN", "CS", "NSM", "BN", "B", "S", "WS",
     "ON", "LRI", "RLI", "FSI", "PDI"
   ]
 
+proc categoryMap(s: string): int =
+  case s
+  of "Lm":
+    ctgLm
+  of "Lo":
+    ctgLo
+  of "Lu":
+    ctgLu
+  of "Ll":
+    ctgLl
+  of "Lt":
+    ctgLt
+  of "Mn":
+    ctgMn
+  of "Mc":
+    ctgMc
+  of "Me":
+    ctgMe
+  of "Nd":
+    ctgNd
+  of "Nl":
+    ctgNl
+  of "No":
+    ctgNo
+  of "Zs":
+    ctgZs
+  of "Zl":
+    ctgZl
+  of "Zp":
+    ctgZp
+  of "Cc":
+    ctgCc
+  of "Cf":
+    ctgCf
+  of "Cs":
+    ctgCs
+  of "Co":
+    ctgCo
+  of "Cn":
+    ctgCn
+  of "Pc":
+    ctgPc
+  of "Pd":
+    ctgPd
+  of "Ps":
+    ctgPs
+  of "Pe":
+    ctgPe
+  of "Pi":
+    ctgPi
+  of "Pf":
+    ctgPf
+  of "Po":
+    ctgPo
+  of "Sm":
+    ctgSm
+  of "Sc":
+    ctgSc
+  of "Sk":
+    ctgSk
+  of "So":
+    ctgSo
+  else:
+    assert false
+    -1
+
 proc parseProps(propsRaw: seq[seq[string]]): seq[seq[int]] =
   result = newSeq[seq[int]](len(propsRaw))
   for i in 0 ..< len(propsRaw):
-    result[i] = @[0, 0]
+    result[i] = @["Cn".categoryMap(), 0]
 
   for cp, props in pairs(propsRaw):
     if isNil(props):
       continue
-    result[cp][Props.CAT.ord] = categoryNames.find(props[0])
+    result[cp][Props.CAT.ord] = props[0].categoryMap()
     result[cp][Props.CCC.ord] = parseInt(props[1])
     assert result[cp][Props.CAT.ord] >= 0
 
@@ -111,10 +202,44 @@ type
     nfdQcNo = $#
     nfkdQcNo = $#
 
+type
+  UnicodeCategory* = distinct int32
+    ## A type for extracting the category
+    ## value out of the raw properties.
+
 const
-  categoryNames* = [
-    $#
-  ]
+  ctgLm* = $#.UnicodeCategory
+  ctgLo* = $#.UnicodeCategory
+  ctgLu* = $#.UnicodeCategory
+  ctgLl* = $#.UnicodeCategory
+  ctgLt* = $#.UnicodeCategory
+  ctgMn* = $#.UnicodeCategory
+  ctgMc* = $#.UnicodeCategory
+  ctgMe* = $#.UnicodeCategory
+  ctgNd* = $#.UnicodeCategory
+  ctgNl* = $#.UnicodeCategory
+  ctgNo* = $#.UnicodeCategory
+  ctgZs* = $#.UnicodeCategory
+  ctgZl* = $#.UnicodeCategory
+  ctgZp* = $#.UnicodeCategory
+  ctgCc* = $#.UnicodeCategory
+  ctgCf* = $#.UnicodeCategory
+  ctgCs* = $#.UnicodeCategory
+  ctgCo* = $#.UnicodeCategory
+  ctgCn* = $#.UnicodeCategory
+  ctgPc* = $#.UnicodeCategory
+  ctgPd* = $#.UnicodeCategory
+  ctgPs* = $#.UnicodeCategory
+  ctgPe* = $#.UnicodeCategory
+  ctgPi* = $#.UnicodeCategory
+  ctgPf* = $#.UnicodeCategory
+  ctgPo* = $#.UnicodeCategory
+  ctgSm* = $#.UnicodeCategory
+  ctgSc* = $#.UnicodeCategory
+  ctgSk* = $#.UnicodeCategory
+  ctgSo* = $#.UnicodeCategory
+
+const
   bidirectionalNames* = [
     $#
   ]
@@ -149,10 +274,7 @@ when isMainModule:
   var propsGen = newSeq[string](stages.stage3.len)
   for i, p in stages.stage3:
     assert len(p) == propsLen
-    propsGen[i] = "[$#]" % join(p, "'i16, ")
-  var categoryNamesGen = newSeq[string](len(categoryNames))
-  for i, cat in categoryNames:
-    categoryNamesGen[i] = "\"$#\"" % cat
+    propsGen[i] = "[$#]" % join(p, "'i32, ")
   var bidirectionalNamesGen = newSeq[string](len(bidirectionalNames))
   for i, bi in bidirectionalNames:
     bidirectionalNamesGen[i] = "\"$#\"" % bi
@@ -166,7 +288,36 @@ when isMainModule:
       intToStr(NfkcQcMaybeMask),
       intToStr(NfdQcNoMask),
       intToStr(NfkdQcNoMask),
-      join(categoryNamesGen, ",\n    "),
+      $ctgLm,
+      $ctgLo,
+      $ctgLu,
+      $ctgLl,
+      $ctgLt,
+      $ctgMn,
+      $ctgMc,
+      $ctgMe,
+      $ctgNd,
+      $ctgNl,
+      $ctgNo,
+      $ctgZs,
+      $ctgZl,
+      $ctgZp,
+      $ctgCc,
+      $ctgCf,
+      $ctgCs,
+      $ctgCo,
+      $ctgCn,
+      $ctgPc,
+      $ctgPd,
+      $ctgPs,
+      $ctgPe,
+      $ctgPi,
+      $ctgPf,
+      $ctgPo,
+      $ctgSm,
+      $ctgSc,
+      $ctgSk,
+      $ctgSo,
       join(bidirectionalNamesGen, ",\n    "),
       prettyTable(stages.stage1, 15, "'u8"),
       prettyTable(stages.stage2, 15, "'u8"),